pyproject.toml
[tool.poetry]
name = "bing-image-scrapper"
version = "0.0.1"
description = "Python library to download bulk of images form Bing.com."
authors = ["Spritan <proyashpsb@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"
aiohttp = "^3.10.10"
tenacity = "^9.0.0"
tqdm = "^4.67.0"
types-tqdm = "^4.66.0.20240417"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
